requirements.txt
dj-database-url==0.5.0
Django==2.2.9
django-anymail==7.0.0
django-cors-headers==3.2.1
django-enumerify==0.5.2
django-filter==2.2.0
django-markdownx==3.0.1
django-reversion==3.0.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djangorestframework-jwt==1.11.0
django-extensions==2.2.6
djoser==2.0.3
drf-yasg==1.17.0
python-dotenv==0.10.3
psycopg2==2.8.4
Pillow==7.0.0
sentry-sdk==0.14.1
uWSGI==2.0.18
mailchimp3==3.0.12

# cairocffi==1.0.2
# CairoSVG==2.3.0
# cffi==1.12.1
# coverage==4.5.2
# cssselect2==0.2.1
# defusedxml==0.5.0
# graphviz==0.13.2
# html5lib==1.0.1
# jsonfield==2.0.2
# pkg-resources==0.0.0
# pycparser==2.19
# pydotplus==2.0.2
# PyJWT==1.7.1
# Pyphen==0.9.5
# tinycss2==0.6.1
# WeasyPrint==45
# webencodings==0.5.1
